scale/src/main/java/com/strategyobject/substrateclient/scale/readers/BoolReader.java
package com.strategyobject.substrateclient.scale.readers;

import com.google.common.base.Preconditions;
import com.strategyobject.substrateclient.common.io.Streamer;
import com.strategyobject.substrateclient.scale.ScaleReader;
import lombok.NonNull;

import java.io.IOException;
import java.io.InputStream;

public class BoolReader implements ScaleReader<Boolean> {
    @Override
    public Boolean read(@NonNull InputStream stream, ScaleReader<?>... readers) throws IOException {
        Preconditions.checkArgument(readers == null || readers.length == 0);

        return Streamer.readByte(stream) != 0;
    }
}
